checkpoint sdt
audit log updates
working test scripts for sdt
automatic calculation of audit num in bb
two stage audit data posting
change audit_num packing in audit log

+++ enclave/enclavestate.h
@@ -61,7 +61,7 @@ struct AUD
 	
 	struct AL auditlog;
 	int audit_offset;
-	uint64_t audit_version_no;
+	uint32_t audit_version_no;
 };
